--- src/components/Head/index.ts
@@ -1,1 +1,0 @@
-export { HeadCustom as Head } from "./Head";
