+++ src/components/MetaLayout/index.ts
@@ -1,0 +1,1 @@
+export { MetaLayout } from "./MetaLayout";
